fix: use correct project context
Uses the `ProjectContext` set on the `Provider` to collect `env` variables
and other context related values. Also refactors `DetectProjectMetadata` to
the `schema` package to prevent import error cycles.
