+++ internal/schema/project.go
@@ -7,8 +7,11 @@ import (
 	"encoding/base32"
 	"errors"
 	"fmt"
+	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/go-git/go-git/v5"
 
 	"github.com/infracost/infracost/internal/logging"
 	"github.com/infracost/infracost/internal/vcs"
@@ -189,6 +192,62 @@ type ProjectMetadata struct {
 	Policies            Policies           `json:"policies,omitempty"`
 	Providers           []ProviderMetadata `json:"providers,omitempty"`
 	RemoteModuleCalls   []string           `json:"remoteModuleCalls,omitempty"`
+}
+
+// DetectProjectMetadata returns a new ProjectMetadata struct initialized
+// from environment variables and the provided path.
+func DetectProjectMetadata(path string) *ProjectMetadata {
+	vcsSubPath := os.Getenv("INFRACOST_VCS_SUB_PATH")
+	terraformWorkspace := os.Getenv("INFRACOST_TERRAFORM_WORKSPACE")
+
+	if vcsSubPath == "" {
+		vcsSubPath = gitSubPath(path)
+	}
+
+	return &ProjectMetadata{
+		Path:               path,
+		TerraformWorkspace: terraformWorkspace,
+		VCSSubPath:         vcsSubPath,
+	}
+}
+
+func gitSubPath(path string) string {
+	topLevel, err := gitToplevel(path)
+	if err != nil {
+		logging.Logger.Debug().Err(err).Msgf("Could not get git top level directory for %s", path)
+		return ""
+	}
+
+	absPath, err := filepath.Abs(path)
+	if err != nil {
+		logging.Logger.Debug().Err(err).Msgf("Could not get absolute path for %s", path)
+		return ""
+	}
+
+	subPath, err := filepath.Rel(topLevel, absPath)
+	if err != nil {
+		logging.Logger.Debug().Err(err).Msgf("Could not get relative path for %s from %s", absPath, topLevel)
+		return ""
+	}
+
+	if subPath == "." {
+		return ""
+	}
+
+	return subPath
+}
+
+func gitToplevel(path string) (string, error) {
+	r, err := git.PlainOpenWithOptions(path, &git.PlainOpenOptions{DetectDotGit: true})
+	if err != nil {
+		return "", fmt.Errorf("failed to detect a git directory in path %s of any of its parent dirs %w", path, err)
+	}
+	wt, err := r.Worktree()
+	if err != nil {
+		return "", fmt.Errorf("failed to return worktree for path %s %w", path, err)
+	}
+
+	return wt.Filesystem.Root(), nil
 }
 
 type ProviderMetadata struct {
